Focus on search bar on launch
Sorry I'm not a programmer so sorry if this is a 'hacky' way to do things. I wanted to make it easier to use the app, every time I launched it I needed to tap the search bar (one extra step), instead of just going ahead and searching for a word.